refactor: answers 에서 deleteAll 구현

+++ src/main/java/nextstep/qna/domain/Answers.java
@@ -1,5 +1,8 @@
 package nextstep.qna.domain;
 
+import nextstep.qna.CannotDeleteException;
+import nextstep.users.domain.NsUser;
+
 import java.util.ArrayList;
 import java.util.List;
 
@@ -10,8 +13,12 @@ public class Answers {
         answers.add(answer);
     }
 
-    public void deleteAll() {
-        for (Answer answer : answers) {
+    public void deleteAll(NsUser loginUser) {
+        for (Answer answer : this.answers) {
+            if (!answer.isOwner(loginUser)) {
+                throw new CannotDeleteException("다른 사람이 쓴 답변이 있어 삭제할 수 없습니다.");
+            }
+
             answer.delete();
         }
     }
